Rename sparse_linear
