+++ thinc/layers/sparselinear.pyx
