+++ thinc/tests/unit/test_sparse_linear.py
@@ -22,7 +22,7 @@ def sgd():
 
 @pytest.mark.xfail
 def test_basic(instances, sgd):
-    from thinc.layers.sparse_linear import LinearModel
+    from thinc.layers.sparselinear import LinearModel
 
     X, y = instances
     nr_class = 3
